refactor: create helper file
